refactor: rework item exceptions
