+++ backend/src/main/java/ch/sectioninformatique/template/item/ItemController.java
@@ -2,14 +2,17 @@ package ch.sectioninformatique.template.item;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import ch.sectioninformatique.template.item.ItemExceptions.ItemNotFoundException;
+import ch.sectioninformatique.template.item.ItemExceptions.UnauthorizedItemException;
 
 /**
  * REST controller for managing items in the system.
