+++ backend/src/main/java/ch/sectioninformatique/template/item/ItemExceptionHandler.java
@@ -1,10 +1,12 @@
 package ch.sectioninformatique.template.item;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.http.HttpStatus;
+
+import ch.sectioninformatique.template.item.ItemExceptions.ItemNotFoundException;
 
 /**
  * Global exception handler for item-related exceptions.
